LayoutTests/imported/w3c/web-platform-tests/html/rendering/non-replaced-elements/tables/table-vspace-hspace-s-expected.txt

CONSOLE MESSAGE: line 2659: Error: assert_equals: marginTop expected "0px" but got "25px"
x
x
x

FAIL table vspace hspace (standards mode) Error: assert_equals: marginTop expected "0px" but got "25px"
